Issue #11: Corrected Integer casting of dimensions.

+++ susi/pycuda_simulation/intensity_volume.py
@@ -258,7 +258,7 @@ class IntensityVolume:
                                     fan_parameters[1]))
 
             # Append to image variables (becomes index 1)
-            slice_dim = np.array([coord_w, coord_h, image_num])
+            slice_dim = np.array([coord_w, coord_h, image_num]).astype(np.int32)
             self.image_variables.append(slice_dim)
 
             # Through downsampling, obtain the output image dimensions
